requirements.txt
pyTelegramBotAPI>=4.17.0
python-dotenv>=1.0.1
openai>=1.30.0
groq>=0.9.0
